+++ component/prometheus/exporter/cloudwatch/config_test.go
@@ -9,6 +9,7 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
+var truePtr = true
 var falsePtr = false
 
 const invalidDiscoveryJobType = `
@@ -102,6 +103,71 @@ discovery {
 }
 `
 
+const staticJobNilToZeroConfig = `
+sts_region = "us-east-2"
+debug = true
+static "super_ec2_instance_id" {
+	regions = ["us-east-2"]
+	namespace = "AWS/EC2"
+	dimensions = {
+		"InstanceId" = "i01u29u12ue1u2c",
+	}
+	metric {
+		name = "CPUUsage"
+		statistics = ["Sum", "Average"]
+		period = "1m"
+	}
+	// setting nil_to_zero on the job level
+	nil_to_zero = false
+}
+`
+
+const staticJobNilToZeroMetricConfig = `
+sts_region = "us-east-2"
+debug = true
+static "super_ec2_instance_id" {
+	regions = ["us-east-2"]
+	namespace = "AWS/EC2"
+	dimensions = {
+		"InstanceId" = "i01u29u12ue1u2c",
+	}
+	metric {
+		name = "CPUUsage"
+		statistics = ["Sum", "Average"]
+		period = "1m"
+		// setting nil_to_zero on the metric level
+		nil_to_zero = false
+	}
+}
+`
+
+const discoveryJobNilToZeroConfig = `
+sts_region = "us-east-2"
+debug = true
+discovery_exported_tags = { "ec2" = ["name"] }
+discovery {
+	type = "sqs"
+	regions = ["us-east-2"]
+	search_tags = {
+		"scrape" = "true",
+	}
+	// setting nil_to_zero on the job level
+	nil_to_zero = false
+	metric {
+		name = "NumberOfMessagesSent"
+		statistics = ["Sum", "Average"]
+		period = "1m"
+	}
+	metric {
+		name = "NumberOfMessagesReceived"
+		statistics = ["Sum", "Average"]
+		period = "1m"
+		// setting nil_to_zero on the metric level
+		nil_to_zero = true
+	}
+}
+`
+
 func TestCloudwatchComponentConfig(t *testing.T) {
 	type testcase struct {
 		raw                 string
@@ -150,7 +216,7 @@ func TestCloudwatchComponentConfig(t *testing.T) {
 							Period:                 60,
 							Length:                 60,
 							Delay:                  0,
-							NilToZero:              &nilToZero,
+							NilToZero:              &defaultNilToZero,
 							AddCloudwatchTimestamp: &addCloudwatchTimestamp,
 						}},
 					},
@@ -184,7 +250,7 @@ func TestCloudwatchComponentConfig(t *testing.T) {
 									Period:                 60,
 									Length:                 60,
 									Delay:                  0,
-									NilToZero:              &nilToZero,
+									NilToZero:              &defaultNilToZero,
 									AddCloudwatchTimestamp: &addCloudwatchTimestamp,
 								},
 								{
@@ -193,7 +259,7 @@ func TestCloudwatchComponentConfig(t *testing.T) {
 									Period:                 60,
 									Length:                 60,
 									Delay:                  0,
-									NilToZero:              &nilToZero,
+									NilToZero:              &defaultNilToZero,
 									AddCloudwatchTimestamp: &addCloudwatchTimestamp,
 								},
 							},
@@ -203,7 +269,7 @@ func TestCloudwatchComponentConfig(t *testing.T) {
 								Length:                 0,
 								Delay:                  0,
 								AddCloudwatchTimestamp: &falsePtr,
-								NilToZero:              &nilToZero,
+								NilToZero:              &defaultNilToZero,
 							},
 						},
 						{
@@ -221,7 +287,7 @@ func TestCloudwatchComponentConfig(t *testing.T) {
 									Period:                 60,
 									Length:                 60,
 									Delay:                  0,
-									NilToZero:              &nilToZero,
+									NilToZero:              &defaultNilToZero,
 									AddCloudwatchTimestamp: &addCloudwatchTimestamp,
 								},
 							},
@@ -231,7 +297,7 @@ func TestCloudwatchComponentConfig(t *testing.T) {
 								Length:                 0,
 								Delay:                  0,
 								AddCloudwatchTimestamp: &falsePtr,
-								NilToZero:              &nilToZero,
+								NilToZero:              &defaultNilToZero,
 							},
 						},
 						{
@@ -250,7 +316,7 @@ func TestCloudwatchComponentConfig(t *testing.T) {
 									Period:                 60,
 									Length:                 3600,
 									Delay:                  0,
-									NilToZero:              &nilToZero,
+									NilToZero:              &defaultNilToZero,
 									AddCloudwatchTimestamp: &addCloudwatchTimestamp,
 								},
 							},
@@ -260,7 +326,128 @@ func TestCloudwatchComponentConfig(t *testing.T) {
 								Length:                 0,
 								Delay:                  0,
 								AddCloudwatchTimestamp: &falsePtr,
-								NilToZero:              &nilToZero,
+								NilToZero:              &defaultNilToZero,
+							},
+						},
+					},
+				},
+			},
+		},
+		"static job nil to zero": {
+			raw: staticJobNilToZeroConfig,
+			expected: yaceConf.ScrapeConf{
+				APIVersion: "v1alpha1",
+				StsRegion:  "us-east-2",
+				Discovery:  yaceConf.Discovery{},
+				Static: []*yaceConf.Static{
+					{
+						Name: "super_ec2_instance_id",
+						// assert an empty role is used as default. IMPORTANT since this
+						// is what YACE looks for delegating to the environment role
+						Roles:      []yaceConf.Role{{}},
+						Regions:    []string{"us-east-2"},
+						Namespace:  "AWS/EC2",
+						CustomTags: []yaceModel.Tag{},
+						Dimensions: []yaceConf.Dimension{
+							{
+								Name:  "InstanceId",
+								Value: "i01u29u12ue1u2c",
+							},
+						},
+						Metrics: []*yaceConf.Metric{{
+							Name:                   "CPUUsage",
+							Statistics:             []string{"Sum", "Average"},
+							Period:                 60,
+							Length:                 60,
+							Delay:                  0,
+							NilToZero:              &falsePtr,
+							AddCloudwatchTimestamp: &addCloudwatchTimestamp,
+						}},
+					},
+				},
+			},
+		},
+		"static job nil to zero metric": {
+			raw: staticJobNilToZeroMetricConfig,
+			expected: yaceConf.ScrapeConf{
+				APIVersion: "v1alpha1",
+				StsRegion:  "us-east-2",
+				Discovery:  yaceConf.Discovery{},
+				Static: []*yaceConf.Static{
+					{
+						Name: "super_ec2_instance_id",
+						// assert an empty role is used as default. IMPORTANT since this
+						// is what YACE looks for delegating to the environment role
+						Roles:      []yaceConf.Role{{}},
+						Regions:    []string{"us-east-2"},
+						Namespace:  "AWS/EC2",
+						CustomTags: []yaceModel.Tag{},
+						Dimensions: []yaceConf.Dimension{
+							{
+								Name:  "InstanceId",
+								Value: "i01u29u12ue1u2c",
+							},
+						},
+						Metrics: []*yaceConf.Metric{{
+							Name:                   "CPUUsage",
+							Statistics:             []string{"Sum", "Average"},
+							Period:                 60,
+							Length:                 60,
+							Delay:                  0,
+							NilToZero:              &falsePtr,
+							AddCloudwatchTimestamp: &addCloudwatchTimestamp,
+						}},
+					},
+				},
+			},
+		},
+		"discovery job nil to zero config": {
+			raw: discoveryJobNilToZeroConfig,
+			expected: yaceConf.ScrapeConf{
+				APIVersion: "v1alpha1",
+				StsRegion:  "us-east-2",
+				Discovery: yaceConf.Discovery{
+					ExportedTagsOnMetrics: yaceModel.ExportedTagsOnMetrics{
+						"ec2": []string{"name"},
+					},
+					Jobs: []*yaceConf.Job{
+						{
+							Regions: []string{"us-east-2"},
+							// assert an empty role is used as default. IMPORTANT since this
+							// is what YACE looks for delegating to the environment role
+							Roles: []yaceConf.Role{{}},
+							Type:  "sqs",
+							SearchTags: []yaceModel.Tag{{
+								Key: "scrape", Value: "true",
+							}},
+							CustomTags: []yaceModel.Tag{},
+							Metrics: []*yaceConf.Metric{
+								{
+									Name:                   "NumberOfMessagesSent",
+									Statistics:             []string{"Sum", "Average"},
+									Period:                 60,
+									Length:                 60,
+									Delay:                  0,
+									NilToZero:              &falsePtr,
+									AddCloudwatchTimestamp: &addCloudwatchTimestamp,
+								},
+								{
+									Name:                   "NumberOfMessagesReceived",
+									Statistics:             []string{"Sum", "Average"},
+									Period:                 60,
+									Length:                 60,
+									Delay:                  0,
+									NilToZero:              &truePtr,
+									AddCloudwatchTimestamp: &addCloudwatchTimestamp,
+								},
+							},
+							RoundingPeriod: nil,
+							JobLevelMetricFields: yaceConf.JobLevelMetricFields{
+								Period:                 0,
+								Length:                 0,
+								Delay:                  0,
+								AddCloudwatchTimestamp: &falsePtr,
+								NilToZero:              &falsePtr,
 							},
 						},
 					},